worker-certificate-generator/requirements.txt
pdfkit
jinja2
pika
psycopg2-binary